qpdf: avoid build failure in autotools_preconfigure()
autotools bbclass tries to clean up build area before re-running
configure if the do_configure() function has changed since the previous
run. This is as easy as deleting and recreating ${B} directory when
separate build directory is enabled, but qpdf make system doesn't
know how to do it, so the fallback is to do 'make clean' to try to get
source back to initial state before running configure again.
The catch is that under some circumstanses qpdf Makefile will error
out on this attempt to 'make clean'.

Set CLEANBROKEN flag to indicate the 'make clean' should be skipped.
Re-running configure seems to work fine without it.

Signed-off-by: S. Lockwood-Childs <sjl@vctlabs.com>